imap: allow the server to update message count between searches
Gmail will only return search results with UIDs that the client already
knows exist (usually from selecting ie. EXAMINEing the mailbox). And
they may not send a new EXISTS value in response to a SEARCH command.
This means that even though the mail arrives, the search never shows it,
and the loop hangs.

By issuing a NOOP we encourage the server to tell us about new messages,
which allows them to appear in the results.